Use `Attribute::name` instead of `id2sym`
AR is primarily string keyed, while we're using symbol keys internally.
This leaks in several methods on `AttributeSet`, which are implemented
on top of methods from `Hash`. We're preserving the "original" key that
AR gave us in the attribute's name, so by using that everywhere we were
using `id2sym` before, we can make AR unaware of the difference