Represent that the raw value may or may not be a proc
This is the groundwork for moving `Attribute::UserProvidedDefault` into
Rust. There are two main differences from `FromUser` here. First is that
the value may be a proc, which needs to be called and have its results
memoized. The second is that it may not have an original attribute.

Right now we're just assuming that we never have a proc, I still want to
only produce that variant in user provided defaults. Unlike the Ruby
version of this code, I've opted to have the proc handling in place for
all attribute types. This is mainly in order to keep the code simple.
I've used `Cell` here since we don't generally want to pattern match on
this memoized value, and I didn't want to make a bunch of methods
mutable that didn't already have to be.