Move `UserProvidedDefault` into Rust
Rather than modifying `FromUser`, I've opted to make this a "separate"
source. `UserProvidedDefault` breaks the invariant of always having an
"original" attribute. It will have one if the attribute is backed by a
database column (this is done to ensure that the attr is considered
changed, and that the new defualt value will be persisted). However, if
you do `attribute :foo, default: :bar` on something not backed by a
database column, there is no original attribute we could use (if we try
to use an uninitialized attribute, the field will incorrectly be
considered changed).