Don't clone in `AttributeSet#write_from_user`
In the Active Record test where we assign a value 20,000 times, the
continuous dropping of the original attribute takes a massive amount of
time. This is because we're cloning the value and dropping the original.
We can get rid of the clone (at least when it's accessed from Rust and
not with `[]`), but it's a bit ugly.